Added a new class called Player and integrated with code.
This new class represents the actual players of the game. The Player constructor requires a number (either 1 or 2) which it stores as the field ‘Owner’. Other fields include a list of Units (which gets added to each time a new unit is created) and a list of resources (which at this stage is just a list of Strings as we haven’t looked into creating proper resources yet). When Start() is run, a Player array is created and two new players are added. When MakeMech is called, it takes an argument of either one (this will prob need to be changed).